Refactor and modularize Tibber integration for improved maintainability
- Removed the existing `tibber.rs` file and split the Tibber integration into multiple modules: `api`, `client`, `types`, and `mod.rs` for better organization and clarity.
- Introduced a new `TibberClient` struct in `client.rs` to encapsulate the logic for interacting with the Tibber API, including caching and price decision-making.
- Added `PriceLevel` and `PricePoint` types in `types.rs` to represent pricing data more effectively.
- Implemented shared client logic in `api.rs` to manage Tibber API calls and state across different parts of the application.
- This restructuring enhances code maintainability and aligns with best practices for modular design.
